% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordbyid.R
\name{getAphiaRecordByID}
\alias{getAphiaRecordByID}
\title{Get Aphia record by ID.}
\usage{
getAphiaRecordByID(id, verbose = FALSE)
}
\arguments{
\item{verbose}{}
}
\value{
Aphia record.
}
\description{
Get Aphia record by ID.
}
\examples{
getAphiaRecordByID(141433)
}

