% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{getAphiaRecords}
\alias{getAphiaRecords}
\title{Get Aphia records.}
\usage{
getAphiaRecords(scientificname, like = TRUE, fuzzy = TRUE,
  marine_only = FALSE, verbose = FALSE)
}
\arguments{
\item{scientificname}{Scientific name, multiple names are currently not supported.}

\item{verbose}{}
}
\description{
Get Aphia records.
}
\examples{
getAphiaRecords("Abra alba")
}

