% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{getAphiaRecordsById}
\alias{getAphiaRecordsById}
\title{Get Aphia records by ID.}
\usage{
getAphiaRecordsById(ids, verbose = FALSE)
}
\arguments{
\item{ids}{The ids.}

\item{verbose}{Verbose output.}
}
\value{
A data frame.
}
\description{
Get Aphia records by ID.
}

