% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/northseamacrozoobenthos.R
\docType{data}
\name{northseamacrozoobenthos}
\alias{northseamacrozoobenthos}
\title{Over 1600 taxonomic, vernacular and trivial names of benthic critters from the North Sea}
\format{a vector of class \code{character}.}
\source{
various
}
\usage{
data(northseamacrozoobenthos)
}
\description{
Used for tests and demonstration.
}
\seealso{
\code{\link{wormsbynames}}, \code{\link{wormsbymatchnames}}
}
\author{
Jan M Holstein , 2017-03-05
}
