% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worms_func.R
\name{wormsbyid}
\alias{wormsbyid}
\title{GET AphiaRecordByAphiaID}
\usage{
wormsbyid(x, verbose = TRUE, ids = FALSE, sleep_btw_chunks_in_sec = 0.01)
}
\arguments{
\item{x}{AphiaIDs}

\item{verbose}{be verbose}

\item{ids}{add column "id" and "name" with running id and search names}

\item{sleep_btw_chunks_in_sec}{pause between requests}
}
\value{
a data frame.
}
\description{
takes more than one AphiaID and retrives AphiaRecords from WoRMS
}
\details{
This function will take a integer vector with AphiaIDs, 
retrive AphiaRecords from www.marinespecies.org using the 
GET /AphiaRecordByAphiaID Method described at
http://www.marinespecies.org/rest/.
Results will be output to a data.frame with each row being a record.

For examples, see  \code{\link{wormsaccepted}}
}
