% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{matchAphiaRecordsByNames}
\alias{matchAphiaRecordsByNames}
\title{Match a scientific name.}
\usage{
matchAphiaRecordsByNames(name, marine_only = FALSE, verbose = FALSE)
}
\arguments{
\item{name}{Scientific name, multiple names are currently not supported.}

\item{verbose}{}
}
\value{
Matching Aphia records.
}
\description{
Match a scientific name.
}
\examples{
matchAphiaRecordsByNames("Buccinum fusiforme")
}

